PF Diff cross tests (#2578)
This change adds the framework code for running Diff cross-tests in PF.
It also adds a simple cross-test to show the framework works. I'll add
additional coverage in a follow-up PR.

I've also exposed the Stdout output of both TF Plan and Pulumi Preview,
so that they can be recorded in the cross-test. This will allow us to
use the cross-tests for testing accurate bridge previews.

related to https://github.com/pulumi/pulumi-terraform-bridge/issues/2297