Plugin Framework bridge cross tests (#2592)
This is a test-only PR - it adds cross-tests for Diff for the plugin
framework for the existing behaviour.

Each test defines a set of schemas and a set of scenarios and the full
matrix of schemas X scenarios is run in cross-tests. For each of these
test cases both pulumi and TF are run with the initial input then a
preview is run with the second set of inputs and the outcome of the
preview is compared.

We also record the stdout for both pulumi and TF for manual comparison
of the preview output. The recordings also contain the inputs in order
to make spot-checking them easier.

The full test plan is available in
https://github.com/pulumi/pulumi-terraform-bridge/issues/2597

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2297
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2597