Add opts to SDKv2 Pulcheck and use Parallel in tests which require env (#2573)
This PR adds a variadic opts parameter to the SDKv2 PulCheck, to match
what PF does. This allows us to pass env vars to pulumitest.

~Use pulumitests env vars to pass bridge feature flags.~ PulumTest env
vars do not work for bridge features because the provider is started
in-process instead of by the engine. Instead we specify bridge features
as arguments to `pulcheck`.
This in turn allows us to use t.Parallel for some of our slowest tests
which require env vars.