Add property value transform for removing secrets and outputs (#2656)
This change adds a propertyvalue Transform for removing secrets and
outputs. This will be useful in the detailed diff algorithm where
secrets and outputs are not relevant, since DetailedDiff only returns
paths to properties. We will instead remove all secrets and inputs when
passed to the algorithm so it doesn't have to handle them in the
internals.