Fix PF provider_server detailed diff handling (#2628)
This change fixes an issue with the `provider_server` implementation's
detailed diff handling. Previously passing a detailed diff to it would
result in the previews being deleted. This is tested as part of
https://github.com/pulumi/pulumi-terraform-bridge/pull/2629

The problem was that we were re-calculating the `diffs` and `replaces`
keys for the GRPC Diff protocol in the `provider_server` implementation
but also doing that incorrectly. Instead this change now makes
`provider_server`'s `marshalDiff` just pass through the `diffs` and
`replaces` which we have already calculated in
https://github.com/pulumi/pulumi-terraform-bridge/blob/1d6b032f3e376af4667c6c4d80a65eff072df807/pkg/pf/tfbridge/provider_diff.go#L108-L109

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2620