Pass state back to the engine if Apply encountered an error (#2695)
In the SDKv2 bridge under PlanResourceChange we are not passing any
state we receive during TF Apply back to the engine if we also received
an error. This causes us to incorrectly miss any resources which were
created but encountered errors during the creation process. The engine
should see these as `ResourceInitError`, which allows the engine to
attempt to update the partially created resource on the next `up`.

This PR fixes the issue by passing the state down to the engine in the
case when we receive an error and a non-nil state from TF during Apply.

related to https://github.com/pulumi/pulumi-gcp/issues/2700
related to https://github.com/pulumi/pulumi-aws/issues/4759

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2696