Pass state back to the engine if Apply encountered an error 2 (#2713)
In the SDKv2 bridge under PlanResourceChange we are not passing any
state we receive during TF Apply back to the engine if we also received
an error. This causes us to incorrectly miss any resources which were
created but encountered errors during the creation process. The engine
should see these as ResourceInitError, which allows the engine to
attempt to update the partially created resource on the next up.

This PR fixes the issue by passing the state down to the engine in the
case when we receive an error and a non-nil state from TF during Apply.

This is the second attempt at this. The first was
https://github.com/pulumi/pulumi-terraform-bridge/pull/2695 but was
reverted because it caused a different panic:
https://github.com/pulumi/pulumi-terraform-bridge/issues/2706. We added
a regression test for that in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2710

The reason for that panic was that we were now creating a non-nil
`InstanceState` with a nil `stateValue` which causes the `ID` function
to panic. This PR fixes both issues by not allowing non-nil states with
nil `stateValue`s and by preventing the panic in `ID`.

There was also a bit of fun with go nil interfaces along the way, which
is the reason why `ApplyResourceChange` now returns a
`shim.InstanceState` interface instead of a `*v2InstanceState2`.
Otherwise we end up creating a non-nil interface with a nil value.

related to https://github.com/pulumi/pulumi-gcp/issues/2700
related to https://github.com/pulumi/pulumi-aws/issues/4759

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2696