Autonaming configuration in Configure and Check (#2675)
This PR implements the bridge part of [Autonaming
Configuration](https://github.com/pulumi/pulumi/issues/1518). See
[RFC](https://github.com/pulumi/pulumi/discussions/17592) for the full
design.

In short, [Protobuf definition for autonaming
configuration](https://github.com/pulumi/pulumi/pull/17810) introduced
the protobuf changes required for the provider-side implementation. With
those, a provider can:

- Declare that it supports autonaming configurations with a response
flag in Configure
- Accept two extra properties in CheckRequest: a proposed name and a
mode to apply it

This PR implements autonaming configuration for all bridged providers.
It passes the configuration from CheckRequest all the way down to the
autonaming module. The module itself is now able to understand the modes
and use the proposed name appropriately:

- For `disable` mode, it will disable autonaming and throw an error if
no explicit name was provided by a user
- For `enforce` mode, it will use the proposed name verbatim. Note that
all checks on that name are ignored: this is by design, so that users
would be able to override the provider's settings with `enforce: true`
if they need to
- For `propose` mode, it will try using the proposed name but can also
modify and validate it. The bridge applies transformations (e.g.
lowercasing the name) and checks max length/character set requirements.
See the test cases for details.

No changes are needed to the provider themselves beyond rolling to the
new version of TF bridge once it's published.

The PR is split into three commits: (1) all the manual changes, (2)
mechanical updates of replay tests, (3) end-to-end tests.

Resolves https://github.com/pulumi/pulumi-terraform-bridge/issues/2722