Ensure all sdkv2 test providers have an id and an update method (#2723)
This change tweaks the `EnsureProviderIsValid` method in `pulcheck` to
always give resources an `id` property and ensure they pass the TF
validation with respect to Update methods.

Previously we'd run into trouble with TF validation when a resource
needs an Update and doesn't have one or it doesn't need an Update and
has one. The rules are non-trivial so I've opted to just make all
resources need an Update and ensure they have one.