Refactor SDKv2 non-set detailed diff tests (#2725)
This is a refactors the rest of the SDKv2 detailed diff tests. Instead
of recording them inline, we now run them as cross-tests and record them
in separate files along with the name of the test and the input values.
This mimics what we did for the PF detailed diff tests in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2592

I've gone through the recordings and believe all of them are correct.