Add a fallback for detailed diff replace decisions to ensure detailed diff is presentation-only (#2757)
This change adds a fallback for the detailed diff replace decision. This
ensures that the detailed diff is presentation only.

If we fail to identify the reason for a replace in the detailed diff
calculation we mark it against `__meta`, similar to what we did before:
https://github.com/pulumi/pulumi-terraform-bridge/blob/a952164c556a86f46dac2ac34e915143cfd7abd8/pkg/tfbridge/provider.go#L1262
If we incorrectly identify a non-existent replace we demote it to an
update/create/delete.

This is flagged behind the Accurate Previews flag.

I've stood up
https://github.com/pulumi/pulumi-terraform-bridge/pull/2747 again as it
got accidentally merged into another branch.

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2674
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2726