Add SDKv2 Detailed Diff tests for Computed properties in sets (#2740)
This PR adds tests around Computed properties in Sets for SDKv2, similar
to what we have on the PF side.

set attributes:
- with computed
- with computed and force new

For the block tests each test has a variant with the computed property
unspecified in the program and one where the property is specified in
the user program.
set blocks:
- with computed at the top level
- with computed at the top level and ForceNew at the top level
- with computed at the top level and a nested ForceNew
- with computed at the nested level
- with computed at the nested level and ForceNew at the top level
- with computed at the nested level and a nested ForceNew

Note that all these tests are currently run without Accurate Previews,
because of
https://github.com/pulumi/pulumi-terraform-bridge/issues/2528. I will
enable the Accurate Previews feature flag for these tests along with the
fix for 2528, so that we can evaluate the differences compared to not
running the feature flag.

The output is non-deterministic without Accurate Previews, so the tests
are recorded but skipped.

necessary for
https://github.com/pulumi/pulumi-terraform-bridge/issues/2528