Write _index.md to local docs dir on Any TF Provider (#2717)
Update: Rebased against
https://github.com/pulumi/pulumi-terraform-bridge/pull/2742 which adds
Cobra. Blocked until that merges.

This pull request adds a `registryDocs` Generator to the initialSetup
call for a dynamic provider.

Adjustments to accommodate this change:

- Because the generation instructions will necessarily include the
string `terraform`, the section needs to be added after the edit rules
are applied.
- The logic to determine which installation instruction to generate is
hinged on `info.Repository`, which is the source repo of the provider.
If it doesn't contain "pulumi" then we're dealing with an Any TF
provider.
- `_index.md` is written to a user-specified (via flag)
`--indexDocOutDir`. This allows us to write to any location in a file
system. Leaving this blank prompts the file to only get written to the
virtual filesystem in the bridge.
- An integration test is added that verifies generation and translation
of the `_index.md` file` from hashicorp/random.

---------

Co-authored-by: Ian Wahbe <me@iwahbe.com>