Improve the error messages in dynamic/internal/shim/grpcutil/replay
This was motivated while trying to add a test for
https://github.com/pulumi/pulumi-terraform-bridge/pull/2815. Beyond adding comments, the
major change is how the failure for differences in nested fields looks. It used to look
like this:

```console
go test -test.run \^TestConfigInDestroy\$
panic: fatal: An assertion has failed: field config does not match logged field \
	&{{} [] [] 0x14000410008} != &{{} [] [] 0x14000410008}
```

Observe that both sides of the `!=` are identical.

It now looks like this:

```console
$ go test -test.run \^TestConfigInDestroy\$
panic: fatal: An assertion has failed: field "terraformv6_pulumi config msgpack" does not match logged field \
	[134 83 71 ...] != [134 82 69 ...]
```

The path to the nested field is specified (`"terraformv6_pulumi config msgpack"`) and we
see a difference in the value displayed.