Refactor PF detailed diff tests (#2814)
This PR refactors the PF detailed diff tests to make test sharding
easier. The main problem was that TestDetailedDiffSet was too large.

This PR:
- Splits up `TestDetailedDiffSet` into a few separate tests. This will
allow the tests to be run on different shards.
- Moves around some of the shared code into the `value_makers.go` file,
instead of duplicating in each test.
- Renames the tests to specify that they are for PF and renamed some of
the utility functions to clarify what they do.


[dbc1c2d](https://github.com/pulumi/pulumi-terraform-bridge/pull/2814/commits/dbc1c2dc6d15b12d93c6821061df6e350b1f6dd0)
has test recordings, the rest are the changes. The test recording
changes are all no-ops related to the change in the struct being
recorded.