Refactor sdkv2 detailed diff tests (#2816)
This PR refactors the SDKv2 detailed diff tests to make test sharding
easier. The main problem was that TestDetailedDiffSet was too large.
Very similar to what
https://github.com/pulumi/pulumi-terraform-bridge/pull/2814 does for PF.

This PR:
- Moves the SDKv2 detailed diff tests to their own folder. This should
make sharing code between them easier without exposing it to the other
tests.
- Splits up `TestDetailedDiffSet` into a few separate tests. This will
allow the tests to be run on different shards.
- Moves around some of the shared code into the `value_makers.go` file,
instead of duplicating in each test.
- Renames the tests to specify that they are for SDKv2 and renamed some
of the utility functions to clarify what they do.


[2c18440](https://github.com/pulumi/pulumi-terraform-bridge/pull/2816/commits/2c18440f01c2aacf598a3fba1e7c2d15ec8e3b8d)
has test recordings, the rest are the changes. The test recording
changes are all no-ops related to the change in the struct being
recorded.