Add code for computing optional sequence diffs (#2865)
Originating from t0yv0/godifft - this may be useful for computing list
diffs for display. Compared to the https://github.com/t0yv0/godifft
version adds rapid tests and a function to apply the edit sequence.

Note that memory is still O(N^2) for large strings, and the naive matrix
implementation may allocate a lot of memory. This is intended for
comparing fairly short strings <1000 in length.