Write a new schema only if necessary
As I explain in my change:

```
For the Schema language in particular, we only write files if the write
would cause a change. This allows build systems like Make to correctly
include bridge-metadata.json and schema.json as build dependencies
without always trying to rebuild.

We don't do this for all systems because reading every file before writing it is
expensive.
```

I've been playing around with using real file dependencies for Go and Make, and the way we
always write bridge metadata (which is then a build input via go:embed) prevents Make from
ever observing that a provider is already built.