Refactor unknown collection supported flag (#2896)
### Context
We have a mechanism for distinguishing between the SDKv1 and the SDKv2
on if they support unknowns for collections. The SDKv1 fails when one is
presented while the SDKv2 works fine

This is `unknownCollectionsSupported` and was introduced in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2061

### Change
This PR is a pure refactor. The flag was introduced on the provider
object but is actually a property of the Schema. This PR changes the
mechanism to be a new shim.Schema interface -
`SchemaWithUnknownCollectionSupported`. The SDKv2 Schema implements this
and the interface implementation can be used directly to infer support
for unknown collections.

This allows us to delete a bunch of code and simplify the
implementation. It also simplifies the implementation in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2891 which is
stacked on top of this.