Fix Accurate Previews for set-nested sets (#2891)
The detailed diff v2 code currently fails to calculate previews for
set-nested sets. This results in a lack of preview for some resources.

The problem is triggered by a panic which we handle in the set element
hashing code. It is caused by a difference in the way we present values
for use by the TF hashing code.

TF expects the values to be the golang-equivalent of the `schema`
values, except for schema.Set:
- `string` for `schema.String`
- `[]intreface{}` for `schema.List`
- `map[string]interface{}` for `schema.Map`
**However:**
- !**`schema.Set` for `schema.Set`**!

Reference:
https://github.com/hashicorp/terraform-plugin-sdk/blob/4d5e3b9a3211f139b41e78ea9ee880c8bf9bb87f/helper/schema/serialize.go#L72

Instead we were using the `makeTerraformInput` code which converts
`resource.PropertyValue` to golang `interface{}`.

This PR adds an option in the conversion context `UseTFSetTypes` which
is only used by the detailed diff code for converting list types where
the TF schema points to a Set into a `shim.Set`. This is only
implemented for schemas which implement the `SchemaWithNewSet`
interface, which are the SDKv1 and SDKv2.

I've also added additional cross-tests around the problematic schema.
Recordings in
[d099de2](https://github.com/pulumi/pulumi-terraform-bridge/pull/2891/commits/d099de25774846fa310b2bf6565accdee45db60d)
 
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2892