Bump `pulumi/pulumi` to 3.152.0
This change bumps the version of `pulumi/pulumi` used to the latest
release, 3.152.0. This is mostly a mechanical change, but the main
motivation is that the `convert.Mapper` interface we rely on for
generating mappings from Terraform to Pulumi now supports parameterized
hints, which are essential in the face of dynamically bridged providers.
See pulumi/pulumi#18671 for more context.