chore: fix downstream_checks.py (#2918)
Relates to https://github.com/pulumi/pulumi-terraform-bridge/issues/2917

The script was confusing to me. It printed SKIPPED for jobs that failed.
It is too easy to miss failure.

It was writing MISSING for cases where some jobs may still be in
progress.

It also tried to handle providers such as awsx that are not bridged -
best exclude these.