PluginFramework encoders now parse "" as nil numbers
This is a workaround until #1667 can guarantee that RawState is parsed correctly and passed to
migration machinery.

In theory there should be no harm to be a little tolerant when parsing incorrectly typed data, but
there may be other use cases that do not agree that "" should be interpreted as nil.

For the moment this workaround fixes a popular issue in `pulumi-aws`:

    https://github.com/pulumi/pulumi-aws/issues/5222