Tolerate Configure panics if one provider errors out (#2931)
Minimally scoped panic recovery to address the scenario found in
https://github.com/pulumi/pulumi-gcp/issues/3007

GCP expects failure in Configure for SDKv2 to be final. When our code
attempts to call Configure for PF after that failure, it results in a
panic.

I do not know if fail-on-first-error is the desired behavior as it seems
preferable to collect all the errors to show to the user if possible.
Therefore this PR is scoped to tolerate panics if there have been
preceding errors.