Fix conversion handling of map type resource elems in schema.go (#2954)
This PR fixes handling of Map-typed Schemas with a Resource Elem in
schema.go. This schema combination is invalid in the SDKv2 but is
produced by some utilities in the bridge. We incorrectly nesting these
values under a list for some reason when converting to the intermediate
golang-type layer.

These schemas are produced to refer to the elements of a Resource Elem
List or Set-type. There is no valid way to refer to the schemas of these
types, so we output a Map-schema with a Resource Elem. I believe this is
currently the only way to produce a Map type with a Resource Elem in the
SDKv2 bridge.

The reason is an odd workaround which was added in
https://github.com/pulumi/pulumi-terraform-bridge/issues/1830 (a
refactor?). There isn't much context around it but it's certainly wrong.
I suspect it was added to work around another bridge bug which was since
fixed and this has remained.

This is part of
https://github.com/pulumi/pulumi-terraform-bridge/issues/2920

The only other alternative here was pile on even more workarounds for
this behaviour. I opted to clean it up instead.