test: add more state upgrade tests (#2957)
Some of these are waiting on #1667 to pass. TestUpgrade_StateUpgraders
is passing though. This happens due to a workaround in place, where
recovering cty.Value fails to match the type but falls back on
approximate methods that seem to give a correct result in simple cases
as this.


https://github.com/pulumi/pulumi-terraform-bridge/blob/ffbd9a5c91c054e49d06bea31e56515edd13f278/pkg/tfshim/sdk-v2/provider.go#L174

Probably good to have this behavior pinned in a test.