chore: remove genrandom test (#2966)
TestGenRandom was covering basic PF resource lifecycle. It was brittle,
written by producing logs and replaying them, and thus susceptible to
bit-rot when the logs changed accidentally. It also depends on a custom
build of a random provider, causing complications for building the
project. Suggest removing the test. To retain coverage, a simple test is
added that does the same lifecycle against the current Pulumi CLI but
has fewer asserts and is less susceptible to bit-rot.