chore: remove difficult to maintain test TestGenUpdates (#2967)
TestUpdateProgram is already running, but TestGenUpdates was reusing an
old version of its logs to replay against the provider one more time.
This perhaps historically made sense allowing to reuse the integration
test as "unit" test, but given our current strategy on relying more on
full-on tests involving Pulumi CLI as the engine, that is redundant for
any test assurances. Suggest removing it. The replay nature of the test
was creating friction when unrelated changes were going in.