PF State Upgrade integration tests (#2968)
This adds schema upgrade test capabilities to the PF Diff cross-test
framework.

This PR also includes two integration tests for the type of schema
migrations we need in the new Cloudflare major version, which are
currently broken: https://github.com/pulumi/pulumi-cloudflare/pull/1057

Note the tests are currently skipped as they do not work in Pulumi.