chore: shim over cty.Value and tftypes.Value representations (#2986)
As part of work on #1667 it is helpful to write code that can operate on
both cty.Value native to SDKv2 solutions and tftypes.Value native to the
Plugin Framework and dynamic bridge solutions. One helpful invariant is
that tftypes.Value does not reorder Set elements like cty.Value does
(canonicalizing the order), and for #1667 it is essential that the order
of set elements is preserved as much as possible. Doing so via a shim
layer makes it easier to accomplish by avoiding intermediate cty.Value
to tftypes.Value conversions. This PR simply adds the new package
without linking it in, for easier review.