testing: Plugin Framework version of upgrade cross tests (#3016)
Ports cross-tests for state upgrade functionality from SDKv2 to PF. As
reflected in TODOs, PF behaves more poorly than the SDKv2 bridge with
respect to unexpected updates caused by Pulumi-side property renames.
This will be resolved with #1667 .

Tests pertaining to MaxItems=1 changes initiated upstream are not ported
because they are less prevalent in PF, we detect MaxItems=1 indirectly
here:
https://github.com/pulumi/pulumi-terraform-bridge/blob/ca70d86e9bd14b8fac8edb74fe119dbcfbda1f23/pkg/pf/internal/pfutils/block.go#L83
but it is somewhat niche.

Fixes: https://github.com/pulumi/pulumi-terraform-bridge/issues/1736 

In particular 1736 is done modulo randomized testing to cover schema
combinations which we can revisit separately. At least for "classical"
tests there are test matrices in place now to cross-test and explore a
particular state transition of interest.

On top of https://github.com/pulumi/pulumi-terraform-bridge/pull/3021