chore: do not fail-fast when testing (#3026)
When making some of the more extensive changes to the bridge, it is
helpful to have a list of all failing tests. Running all the tests
locally takes a fair bit of time so currently in practice we often rely
on CI to identify test failures. It is helpful to identify all of them
instead of just the first one.

Relates: https://github.com/pulumi/pulumi-terraform-bridge/issues/3015