Preserve TF raw state as deltas in Pulumi state (#2945)
This change fixes a long-standing problem of resource state data being
distorted by the bridge, which is particularly sensitive around provider
upgrade scenarios where schemas are changing and the TF provider expects
to receive the older state record verbatim to auto-migrate it to the new
schema. Before the change states were stored as PropertyMap and
converted to TF representation on-read against the currently known
provider schema, which may be different from the schema at the write
time. The fix is two-fold:

- at write time, compute the delta between TF raw state and Pulumi
PropertyMap and store it in the PropertyMap
- at read time, if the delta is found, reconstruct the original
byte-for-byte Terraform RawState and pass it to Terraform to migrate.

The change is feature-flagged and we aim for a gradual rollout but
eventually make the new behavior the default.

- Set EnableRawStateDelta: true on info.Provider to enable for a
provider
- Set PULUMI_RAW_STATE_DELTA_ENABLED=true to dynamically enable (or
disable) this behavior

One particular technical complication is around Set structures, which
Pulumi does not model. For the deltas to work correctly the set element
ordering in Pulumi array representation needs to be consistent between
the TF and Pulumi states. This is accomplished for the SDKv2 bridge by
preserving the cty.Value-canonical ordering but for the PF bridge by
preserving the "pulumi-natural" ordering of set elements.

Fixes #1667