Fix PF WriteOnly attributes skip (#3043)
~This is a pure refactor. It removes a redundant internal interface from
the PF bridge.~

This started as a refactor but is actually a bug fix. Looks like we did
not implement the PF portion of
https://github.com/pulumi/pulumi-terraform-bridge/pull/2933 correctly.
We also had a test which asserted on the wrong behaviour.

This PR fixes skipping PF WriteOnly attributes during tfgen. It also
simplifies the PF code by removing a redundant interface.

`AttrLike` is a copy of the PF `fwschema..Attribute` which has
`WriteOnly` implemented. This makes the method non-optional and adding
it to an internal interface is not a breaking change.