Fix required properties with default in TF projecting as non-optional in Pulumi (#3035)
This PR fixes the schema generation for TF properties which are marked
as `Required` and have a default.

In TF these become optional, as the default will get applied. This PR
makes this also the case in Pulumi. SDKs should no longer require such
properties to be specified as inputs.

This change should be fully backwards-compatible, see
[here](https://github.com/pulumi/pulumi-terraform-bridge/issues/2204#issuecomment-2842609784)
for a per-language discussion.
 
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2204