Automate pu/pu version bump in tests (#3052)
We pinned the pu/pu verison used for tests instead of having it floating
to reduce breaking change impact.

This PR automates bumping the versions and adds a check that the version
used in tests matches the one in go.mod.
It also bumps the version for pu/pu used in tests and re-records tests
whose output has changed.



Example failure of the check here:
https://github.com/pulumi/pulumi-terraform-bridge/actions/runs/14884095956/job/41798968807