Upgrade pulumi to v3.167.0 (#3054)
Also upgrade github.com/pulumi/pulumi-java/pkg v1.11.0 as that is
required to be build-compatible with latest Pulumi.

Changes around `pschema.BindSpec` options are now satisifed by opting
into AllowDanglingReferences: true which I believe was the prior
behavior. We need https://github.com/pulumi/pulumi-aws/issues/5250 and
similar to be solved in providers before we can tighten this to the
strict mode.