Fix RawStateDelta computation for overflowing big ints and floats
The test suite in pulumi-gcp points out a corner case that was not handled previously and may lead to:

    error: recovered raw state does not byte-for-byte match the original raw state

This is the case of overflowing numbers (ints and floats) that cannot fit precisely in the float64 representation used
by Pulumi. If the bridged provider maintainer overrides the bridge metadata to project these numbers as strings in the
SDKs, everything works as expected. However there are practical cases where this maintainer action has not been taken
yet; one reason to defer it is that changing the type from Number to String is a breaking change on the SDKs and may
need to await a major version.

The change tightens equality to be stricter than equal-up-to-float64 precision, and if Pulumi representation does not
equal the TF representation exactly, the provider will now backtrack and write the exact TF representation into the Raw
State Delta, instructing the reader to replace the Pulumi representation with the TF representation.

See also: https://github.com/pulumi/pulumi-gcp/pull/3212