PF Create cross-tests (#3065)
This PR adds cross-test for `Create` for the PF bridge. These tests run
a given resource configuration for a given schema under both Pulumi and
TF and compare the inputs which the underlying TF provider receive.

This tests the coversion pipeline for input values which we pass to the
TF provider. Similar to the `Create` cross-test in the sdkv2 bridge:
https://github.com/pulumi/pulumi-terraform-bridge/blob/80add2e0cc358683693e19236cc5ad874ae0d624/pkg/internal/tests/cross-tests/create.go#L54