Fix PF Dynamic types (#3066)
This PR fixes the PF bridge support for dynamic types when the attribute
is not specified. Specifically, previously it'd default to an empty
object and now the default is a nil PseudoDynamicType.

This is supported by a cross-tests which compares against TF behaviour
which does the same.

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/3067