Migrate golangci lint (#3097)
Upgrade the linter to V2 and fix various issues discovered by the new
version.

V1 has some non-determinism with its lints causing PRs to pass the
linter and then it fails on master, for example:
https://github.com/pulumi/pulumi-terraform-bridge/issues/3096

The changes are mostly making the imports consistent and removing
double-imports. I've also used the chance to add a rule to import some
common libraries with the same alias.