Fix PF dynamic type diff error (#3098)
In https://github.com/pulumi/pulumi-terraform-bridge/pull/2945/ we
accidentally changed the behaviour for providers where RawStateDeltas
were not enabled. After 2945 we would unconditionally produce the JSON
serialization of the state even if it isn't used.

It is only used for upgrading state and for RawStateDeltas. The
unintended effect here is that we exposed
https://github.com/pulumi/pulumi-terraform-bridge/issues/3078 to all
users of attributes of DynamicType even if no upgrades or RawStateDeltas
are used.

This PR changes that to not produce the JSON serialization if not
necessary. I've also added some tests for both the behaviour fixed here
and the behaviour which needs
https://github.com/pulumi/pulumi-terraform-bridge/issues/3078

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/3095