Add NoDanglingReferences option (#3118)
Historically providers such as pulumi-aws generated schemata that had
type references that cannot resolve, and relied
on the quirks of the Node SDK generator to make use of these references
to link in overlay types.

This is going away in upcoming major release of pulumi-aws, and newer
providers should be strict about generating the
schema.

When the new flag NoDanglingReferences is set to `true`, bridge will
perform strict schema validity checks before
emitting a schema during `make tfgen`.

Fixes #3116