Fix diffing assets and archives (#3115)
This PR fixes the diffing of resource properties which contain Pulumi
Assets or Archives. We were not correctly handling the assets coming
from the old properties both for with RawStateDeltas enabled and
without. Instead we were using the new property assets in place of the
old ones and that produced no diff for these properties.

This fixes both:
- Without raw state deltas, the asset table is parsed as part of
transforming the property values to the TF domain - we can use these in
the diff algorithm.
- With raw state deltas, we now add a recursive pass over the old
property values, which produces an asset table.

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/3102