Remove ID attribute from schema in test providers (#3121)
This PR removes the mandatory addition of an "id" property to all TF
schemas created by the `pulcheck` test library. That is not necessary in
TF so we should not be requiring it in our tests.

It also adds some additional checks for some of our tests.

This is a test-only change.