Default to failing build on extra mapping error (#3144)
This pull request removes the `PULUMI_EXTRA_MAPPING_ERROR` environment
variable from the pulumi-terraform-bridge.

Of note: an "extra mapping" is when the Pulumi provider info attempts to
map a resource or data source that does not exist in the upstream
provider. A "missing mapping" is when a TF resource is not mapped to a
Pulumi resource. This pull request ONLY deals with the former.

Previous work on this was partially implemented this
https://github.com/pulumi/pulumi-terraform-bridge/pull/616/files only
set this error by default in resource generation, but not data source
generation. This pull request fixes that.

Additionally, extra field mappings are [now caught in
Validate](https://github.com/pulumi/pulumi-terraform-bridge/blob/e774624b0fd4804dcfce8ba37debbba1bcdad16d/pkg/tfbridge/info/validate.go#L85)
so we no longer need to check for them during generation.

This pull request adds tests to verify each scenario - extra mappings
should trigger errors.

It is my opinion that this is a safe clean up - as stated, this has been
the default for Resources for a few years now, and we have [been setting
PULUMI_EXTRA_MAPPING_ERROR to `true` in
upgrade-provider](https://github.com/pulumi/upgrade-provider/blob/main/upgrade/upgrade_provider.go#L50)
so we should see no churn.