Add --resources flag to filter dynamic provider resources
This commit introduces a --resources flag that allows users to specify
which resources should be included when generating a dynamic Pulumi
provider from a Terraform provider. This enables creating focused
providers that contain only the subset of resources needed for a
particular use case.

The implementation adds a Resources field to both Args and Value
structs to track the list of resource names to include. When specified,
the filtering is applied in the providerInfo function by leveraging
the existing IgnoreMappings mechanism - all resources not in the
include list are added to IgnoreMappings, effectively excluding them
from the generated provider.

Users can now run commands like:

  pulumi package add terraform aws --resources aws_instance,aws_vpc

This results in a focused AWS provider containing only the specified
resources, reducing the generated provider size and improving
development ergonomics when working with large Terraform providers.

Fixes https://github.com/pulumi/pulumi-terraform-provider/issues/91